requirements.txt
Flask==1.1.4
Flask_RESTful==0.3.10
openai==0.27.4
python-dotenv==1.0.0
